requirements.txt
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
albumentations==1.4.8
tqdm==4.66.4
matplotlib==3.9.0
numpy==1.26.4
Pillow==9.5.0








